% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoGeneRankerFunctions.R
\name{addParameters}
\alias{addParameters}
\title{This method allows you to import all the parameters that you need for use with ease of use for the other methods.}
\usage{
addParameters(r, delta, zeta, tau, phi, lambda, eta, weight, LG, LC)
}
\arguments{
\item{r}{This is the global restart probability. It has a range of 0 to 1.}

\item{delta}{This is the probability of jumping between layers. It has a range of 0 to 1.}

\item{tau}{This is the Gene Layers Restart Probability. It is a 3 digit vector for (PPI, PATH, COEX) that adds up to 1.}

\item{phi}{This is the Phenotype Layers Restart Probability. It si a 3 digit cevtor for (Genotype, EL, LTSS) that adds up to 1.}

\item{lambda}{This is the Inter Network Jump Probability. It has a range of 0 to 1.}

\item{eta}{This is the networks restart probability. It has a range from 0 to 1.}

\item{weight}{This is the weight given to the matrix. it has a range from 0 to 1.}

\item{LG}{This is the number of genes that will be used.}

\item{LC}{This is the number of phenotypes that will be used.}
}
\value{
This returns a list of the parameters.
}
\description{
This method allows you to import all the parameters that you need for use with ease of use for the other methods.
}
