% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoGeneRankerFunctions.R
\name{calculate.p_values}
\alias{calculate.p_values}
\title{Generate S random seeds as length of GeneSeeds excluding  actual Seeds
 Run the algorithm for N times, store N result vector
  If (the rank of the gene in random seed <= (actual seed rank + Offset) ) then
    it will be counted as 1, otherwise 0}
\usage{
calculate.p_values(RWGeneRanks, Rand_Seed_Gene_Rank, output_dir, no.cores = 15)
}
\arguments{
\item{no.cores}{}
}
\value{

}
\description{
Generate S random seeds as length of GeneSeeds excluding  actual Seeds
 Run the algorithm for N times, store N result vector
  If (the rank of the gene in random seed <= (actual seed rank + Offset) ) then
    it will be counted as 1, otherwise 0
}
