% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoGeneRankerFunctions.R
\name{create.WalkMatrix}
\alias{create.WalkMatrix}
\title{Generates a Walk matrix file from Gene and Cultivar data}
\usage{
create.WalkMatrix(inputFileName, params, numCores)
}
\arguments{
\item{inputFileName}{name of the txt file that contains the gene and cultivar data.}

\item{numCores}{number of cores used for parallel processing.
- recommended to use detectCores() method from the parallel package.}
}
\value{
ID of WalkMatrix File, the RDA file is saved to your files.
}
\description{
Generates a walk matrix file from the data given by the user.
The data for each gene and cultivar is added to supraadjency matrixes,
bipartite matrixes, and suprabipartite matrixes.These are combined to make transition and
transistion multiplex networks. A multiplex heterogeneous network is created by
combining the transition and transition multiplex networks.
That network then gets saved to an RDA file
}
