% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoGeneRankerFunctions.R
\name{generate.random.seed.vector}
\alias{generate.random.seed.vector}
\title{stratified sampling by using connectivity of the matrix
 1. Sort genes by their degree
 2. Divided the Genes into 10 groups
 3. Get the groups of SeedGenes
 4. Random sample from genes from the same degree of genes}
\usage{
generate.random.seed.vector(
  output_dir,
  WM_ID,
  GeneSeeds,
  GeneConnectivityDF,
  CultSeeds,
  CultConnectivityDF,
  S = 1000,
  no.groups.gene = 10,
  no.groups.cult = 5
)
}
\arguments{
\item{output_dir}{File directory to output the data.}

\item{WM_ID}{Id of the walk matrix used.}

\item{no.groups.cult}{}
}
\value{

}
\description{
stratified sampling by using connectivity of the matrix
 1. Sort genes by their degree
 2. Divided the Genes into 10 groups
 3. Get the groups of SeedGenes
 4. Random sample from genes from the same degree of genes
}
