% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoGeneRankerFunctions.R
\name{CreateWalkMatrix}
\alias{CreateWalkMatrix}
\title{Create Walk Matrix}
\usage{
CreateWalkMatrix(
  inputFileName,
  numCores = 1,
  delta = 0.5,
  zeta = 0.5,
  lambda = 0.5
)
}
\arguments{
\item{inputFileName}{The name of the text file that contains the names of}

\item{numCores}{This is the number of cores used for parallel processing.}

\item{delta}{This is the probability of jumping between gene layers. It has a
range of 0 to 1. It has a default value of 0.5.}

\item{zeta}{This is the probability of jumping between phenotype layers. It
has a range of 0 to 1. It has a default value of 0.5.}

\item{lambda}{This is the Inter-Network Jump Probability. It has a range of 0
to 1. It has a default value of 0.5.}
}
\value{
This returns a list that contains a walk matrix, list of sorted genes
  by name, list of phenotypes sorted by name, the degree of all genes in the
  multiplex heterogeneous network, the degree of all phenotypes in the
  network, the number of gene layer files, the number of phenotype layer
  files, the number of gene pool nodes, and the number of phenotype nodes.
}
\description{
Generates a Walk matrix from Gene and Phenotype data.
}
\examples{
\dontrun{
CreateWalkMatrix('myInput.txt')
CreateWalkMatrix('file.txt', detectCores(), 0.4, 0.7, 0.9)
}
}
